pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "serverless-openapi-generator"
version = "1.0.0"
authors = [
  { name="Cline", email="cline@example.com" },
]
description = "A standalone tool to generate OpenAPI schemas from serverless.yml files."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyYAML",
    "requests",
    "referencing",
    "jsonschema-spec",
    "openapi-spec-validator",
]

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-mock>=3.14.1",
    "requests-mock>=1.12.1",
]

[project.scripts]
openapi-gen = "serverless_openapi_generator.openapi_generator:main"
openapi-validate = "openapi_spec_validator.__main__:main"
